Fix return type

+++ aoc2023/day04/Program.cs
@@ -13,7 +13,7 @@ namespace aoc.aoc2023.day04
             Console.WriteLine(Part2(input));
         }
 
-        private static double Part1(int[] input) =>
+        private static int Part1(int[] input) =>
             input.Sum(c => c > 0 ? 1 << (c - 1) : 0);
 
         private static int Part2(int[] input)
